Delete .env
we should to store .env on our school server, but not in GitHub.
The .env file stores environment variables, typically containing sensitive or configurable data such as database credentials, API keys, or application settings. It separates configuration from source code and makes it easier to manage settings for different environments.